# gateway/cmd/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: GatewayOps API
  version: 1.0.0
  description: |
    GatewayOps is an enterprise MCP (Model Context Protocol) Gateway that provides
    secure, observable, and scalable access to MCP servers.

    ## Authentication

    All API requests require authentication using a Bearer token:

    ```
    Authorization: Bearer gwo_prd_...
    ```

    API keys can be created in the dashboard or via the API.

    ## Rate Limiting

    API requests are rate limited based on your API key configuration.
    Rate limit headers are included in all responses:

    - `X-RateLimit-Limit`: Maximum requests per minute
    - `X-RateLimit-Remaining`: Remaining requests in the current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

  contact:
    name: GatewayOps Support
    email: support@gatewayops.com
    url: https://gatewayops.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gatewayops-api.fly.dev
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /v1/roles:
    get:
      tags: [RBAC]
      summary: List roles
      description: List all roles.
      operationId: listRoles
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'

    post:
      tags: [RBAC]
      summary: Create role
      description: Create a new custom role.
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleInput'
      responses:
        '201':
          description: Created role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'

  # Safety
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
...
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
        isBuiltin:
          type: boolean
        createdAt:
          type: string
          format: date-time
    RoleInput:
      type: object
      required: [name, permissions]
      properties:
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
...